TST: Move test of ad grid operators to separate file
These used to be located in test_operators, breaking with the convention
of one test file per module.
