+++ tests/numerics/ad/test_operators.py
@@ -1,20 +1,14 @@
 """Test collection for Ad representations of several operators.
 
 Checks performed include the following:
-    test_elementary_operations: Checks are made to ensure that basic arithmetic operations
-        are performed correctly;
+    test_elementary_operations: Checks are made to ensure that basic arithmetic
+        operations are performed correctly;
     test_copy_operator_tree: Testing of functionality under copy.copy and deepcopy;
     test_elementary_wrappers: Test wrapping of fields (scalars, arrays and matrices);
     test_ad_variable_creation: The generation of variables should ensure that copies and
         newly created variables are returned in the expected manner;
-    test_ad_variable_evaluation: Variable wrappers are tested as expected under evaluation
-    test_time_differentiation: Covers the pp.ad.dt operator;
-    test_subdomain_projections: Operators for restriction and prolongation are checked for
-        both faces and cells;
-    test_mortar_projections: Projections between mortar grids and subdomain grids;
-    test_boundary_grid_projection:  Tests are conducted on the boundary projection
-        operator and its inverse;
-    test_trace and test_divergence: Operators for discrete traces and divergences
+    test_ad_variable_evaluation: Variable wrappers are tested as expected under
+    evaluation test_time_differentiation: Covers the pp.ad.dt operator;
     test_ad_discretization_class: test for AD discretizations;
     test_arithmetic_operations_on_ad_objects: Basic Ad operators combined with standard
         arithmetic operations are tested.
@@ -968,383 +962,6 @@ def geometry_information(
     return n_cells, n_faces, n_mortar_cells
 
 
-@pytest.fixture
-def mdg():
-    """Provide a mixed-dimensional grid for the tests."""
-    fracs = [np.array([[0, 2], [1, 1]]), np.array([[1, 1], [0, 2]])]
-    md_grid = pp.meshing.cart_grid(fracs, np.array([2, 2]))
-    return md_grid
-
-
-@pytest.mark.integtest
-@pytest.mark.parametrize("scalar", [True, False])
-def test_subdomain_projections(mdg, scalar):
-    """Test of subdomain projections. Both face and cell restriction and prolongation.
-
-    Test three specific cases:
-        1. Projections generated by passing a md-grid and a list of grids are identical
-        2. All projections for all grids (individually) in a simple md-grid.
-        3. Combined projections for list of grids.
-
-    """
-    proj_dim = 1 if scalar else mdg.dim_max()
-    n_cells, n_faces, _ = geometry_information(mdg, proj_dim)
-
-    subdomains = mdg.subdomains()
-    proj = pp.ad.SubdomainProjections(subdomains=subdomains, dim=proj_dim)
-
-    cell_start = np.cumsum(
-        np.hstack((0, np.array([sd.num_cells for sd in subdomains])))
-    )
-    face_start = np.cumsum(
-        np.hstack((0, np.array([sd.num_faces for sd in subdomains])))
-    )
-
-    # Helper method to get indices for sparse matrices
-    def _mat_inds(nc, nf, grid_ind, dim, cell_start, face_start):
-        cell_inds = np.arange(cell_start[grid_ind], cell_start[grid_ind + 1])
-        face_inds = np.arange(face_start[grid_ind], face_start[grid_ind + 1])
-
-        data_cell = np.ones(nc * dim)
-        row_cell = np.arange(nc * dim)
-        data_face = np.ones(nf * dim)
-        row_face = np.arange(nf * dim)
-        col_cell = pp.fvutils.expand_indices_nd(cell_inds, dim)
-        col_face = pp.fvutils.expand_indices_nd(face_inds, dim)
-        return row_cell, col_cell, data_cell, row_face, col_face, data_face
-
-    # Test projection of one fracture at a time for the full set of grids
-    for sd in subdomains:
-        ind = _list_ind_of_grid(subdomains, sd)
-
-        nc, nf = sd.num_cells, sd.num_faces
-
-        num_rows_cell = nc * proj_dim
-        num_rows_face = nf * proj_dim
-
-        row_cell, col_cell, data_cell, row_face, col_face, data_face = _mat_inds(
-            nc, nf, ind, proj_dim, cell_start, face_start
-        )
-
-        known_cell_proj = sps.coo_matrix(
-            (data_cell, (row_cell, col_cell)), shape=(num_rows_cell, n_cells)
-        ).tocsr()
-        known_face_proj = sps.coo_matrix(
-            (data_face, (row_face, col_face)), shape=(num_rows_face, n_faces)
-        ).tocsr()
-
-        assert _compare_matrices(proj.cell_restriction([sd]), known_cell_proj)
-        assert _compare_matrices(proj.cell_prolongation([sd]), known_cell_proj.T)
-        assert _compare_matrices(proj.face_restriction([sd]), known_face_proj)
-        assert _compare_matrices(proj.face_prolongation([sd]), known_face_proj.T)
-
-    # Project between the full grid and both 1d grids (to combine two grids)
-    g1, g2 = mdg.subdomains(dim=1)
-    rc1, cc1, dc1, rf1, cf1, df1 = _mat_inds(
-        g1.num_cells,
-        g1.num_faces,
-        _list_ind_of_grid(subdomains, g1),
-        proj_dim,
-        cell_start,
-        face_start,
-    )
-    rc2, cc2, dc2, rf2, cf2, df2 = _mat_inds(
-        g2.num_cells,
-        g2.num_faces,
-        _list_ind_of_grid(subdomains, g2),
-        proj_dim,
-        cell_start,
-        face_start,
-    )
-
-    # Adjust the indices of the second grid, we will stack the matrices.
-    rc2 += rc1.size
-    rf2 += rf1.size
-    num_rows_cell = (g1.num_cells + g2.num_cells) * proj_dim
-    num_rows_face = (g1.num_faces + g2.num_faces) * proj_dim
-
-    known_cell_proj = sps.coo_matrix(
-        (np.hstack((dc1, dc2)), (np.hstack((rc1, rc2)), np.hstack((cc1, cc2)))),
-        shape=(num_rows_cell, n_cells),
-    ).tocsr()
-    known_face_proj = sps.coo_matrix(
-        (np.hstack((df1, df2)), (np.hstack((rf1, rf2)), np.hstack((cf1, cf2)))),
-        shape=(num_rows_face, n_faces),
-    ).tocsr()
-
-    assert _compare_matrices(proj.cell_restriction([g1, g2]), known_cell_proj)
-    assert _compare_matrices(proj.cell_prolongation([g1, g2]), known_cell_proj.T)
-    assert _compare_matrices(proj.face_restriction([g1, g2]), known_face_proj)
-    assert _compare_matrices(proj.face_prolongation([g1, g2]), known_face_proj.T)
-
-
-@pytest.mark.integtest
-@pytest.mark.parametrize("scalar", [True, False])
-def test_mortar_projections(mdg, scalar):
-    # Test of mortar projections between mortar grids and standard subdomain grids.
-
-    # Define the dimension of the field being projected
-    proj_dim = 1 if scalar else mdg.dim_max()
-
-    # Collect geometrical and grid objects
-    n_cells, n_faces, n_mortar_cells = geometry_information(mdg, proj_dim)
-
-    g0 = mdg.subdomains(dim=2)[0]
-    g1, g2 = mdg.subdomains(dim=1)
-    g3 = mdg.subdomains(dim=0)[0]
-
-    intf01 = mdg.subdomain_pair_to_interface((g0, g1))
-    intf02 = mdg.subdomain_pair_to_interface((g0, g2))
-
-    intf13 = mdg.subdomain_pair_to_interface((g1, g3))
-    intf23 = mdg.subdomain_pair_to_interface((g2, g3))
-
-    # Compute reference projection matrices
-    face_start = proj_dim * np.cumsum(
-        np.hstack((0, np.array([g.num_faces for g in mdg.subdomains()])))
-    )
-    cell_start = proj_dim * np.cumsum(
-        np.hstack((0, np.array([g.num_cells for g in mdg.subdomains()])))
-    )
-    mortar_start = proj_dim * np.cumsum(
-        np.hstack((0, np.array([m.num_cells for m in mdg.interfaces()])))
-    )
-
-    f0 = np.hstack(
-        (
-            sparse_array_to_row_col_data(
-                intf01.mortar_to_primary_int(nd=proj_dim), True
-            )[0],
-            sparse_array_to_row_col_data(
-                intf02.mortar_to_primary_int(nd=proj_dim), True
-            )[0],
-        )
-    )
-    f1 = sparse_array_to_row_col_data(intf13.mortar_to_primary_int(nd=proj_dim), True)[
-        0
-    ]
-    f2 = sparse_array_to_row_col_data(intf23.mortar_to_primary_int(nd=proj_dim), True)[
-        0
-    ]
-
-    c0 = sparse_array_to_row_col_data(
-        intf01.mortar_to_secondary_int(nd=proj_dim), True
-    )[0]
-    c1 = sparse_array_to_row_col_data(
-        intf02.mortar_to_secondary_int(nd=proj_dim), True
-    )[0]
-    c2 = np.hstack(
-        (
-            sparse_array_to_row_col_data(
-                intf13.mortar_to_secondary_int(nd=proj_dim), True
-            )[0],
-            sparse_array_to_row_col_data(
-                intf23.mortar_to_secondary_int(nd=proj_dim), True
-            )[0],
-        )
-    )
-
-    m0 = sparse_array_to_row_col_data(
-        intf01.mortar_to_secondary_int(nd=proj_dim), True
-    )[1]
-    m1 = sparse_array_to_row_col_data(
-        intf02.mortar_to_secondary_int(nd=proj_dim), True
-    )[1]
-    m2 = sparse_array_to_row_col_data(
-        intf13.mortar_to_secondary_int(nd=proj_dim), True
-    )[1]
-    m3 = sparse_array_to_row_col_data(
-        intf23.mortar_to_secondary_int(nd=proj_dim), True
-    )[1]
-
-    # collect destination indexes
-    face_ixd = np.hstack(
-        tuple([f_idx + face_start[i] for i, f_idx in enumerate([f0, f1, f2])])
-    )
-    cell_ixd = np.hstack(
-        tuple([c_idx + cell_start[i + 1] for i, c_idx in enumerate([c0, c1, c2])])
-    )
-    mort_ixd = np.hstack(
-        tuple([m_idx + mortar_start[i] for i, m_idx in enumerate([m0, m1, m2, m3])])
-    )
-
-    proj_known_higher = sps.coo_matrix(
-        (np.ones(n_mortar_cells), (face_ixd, np.arange(n_mortar_cells))),
-        shape=(n_faces, n_mortar_cells),
-    ).tocsr()
-
-    proj_known_lower = sps.coo_matrix(
-        (np.ones(n_mortar_cells), (cell_ixd, mort_ixd)), shape=(n_cells, n_mortar_cells)
-    ).tocsr()
-
-    # Also test block matrices for the sign of mortar projections.
-    # This is a diagonal matrix with first -1, then 1.
-    # If this test fails, something is fundamentally wrong.
-    vals = np.array([])
-    for intf in mdg.interfaces():
-        sz = int(np.round(intf.num_cells / 2) * proj_dim)
-        vals = np.hstack((vals, -np.ones(sz), np.ones(sz)))
-
-    known_sgn_mat = sps.dia_matrix((vals, 0), shape=(n_mortar_cells, n_mortar_cells))
-
-    # Compute the object being tested
-    proj = pp.ad.MortarProjections(
-        subdomains=mdg.subdomains(), interfaces=mdg.interfaces(), mdg=mdg, dim=proj_dim
-    )
-
-    assert _compare_matrices(proj_known_higher, proj.mortar_to_primary_int)
-    assert _compare_matrices(proj_known_higher, proj.mortar_to_primary_avg)
-    assert _compare_matrices(proj_known_higher.T, proj.primary_to_mortar_int)
-    assert _compare_matrices(proj_known_higher.T, proj.primary_to_mortar_avg)
-    assert _compare_matrices(proj_known_lower, proj.mortar_to_secondary_int)
-    assert _compare_matrices(known_sgn_mat, proj.sign_of_mortar_sides)
-
-
-@pytest.mark.integtest
-@pytest.mark.parametrize("scalar", [True, False])
-def test_boundary_grid_projection(mdg: pp.MixedDimensionalGrid, scalar: bool):
-    """Three main functionalities being tested:
-    1) That we can create a boundary projection operator with the correct size and items.
-    2) Specifically that the top-dimensional grid and one of the fracture grids
-       contribute to the boundary projection operator, while the third has a projection
-       matrix with zero rows.
-    3) Projection from a subdomain to a boundary is consistent with its reverse.
-
-    """
-    proj_dim = 1 if scalar else mdg.dim_max()
-    _, num_faces, _ = geometry_information(mdg, proj_dim)
-    num_cells = sum([bg.num_cells for bg in mdg.boundaries()]) * proj_dim
-
-    g_0 = mdg.subdomains(dim=2)[0]
-    g_1, g_2 = mdg.subdomains(dim=1)
-    # Compute geometry for the mixed-dimensional grid. This is needed for
-    # boundary projection operator.
-    mdg.compute_geometry()
-    projection = pp.ad.BoundaryProjection(mdg, mdg.subdomains(), proj_dim)
-    # Obtaining sparse matrices from the AD Operators.
-    subdomain_to_boundary = projection.subdomain_to_boundary.parse(mdg)
-    boundary_to_subdomain = projection.boundary_to_subdomain.parse(mdg)
-    # Check sizes.
-    assert subdomain_to_boundary.shape == (num_cells, num_faces)
-    assert boundary_to_subdomain.shape == (num_faces, num_cells)
-
-    # Check that the projection matrix for the top-dimensional grid is non-zero.
-    # The matrix has eight boundary faces.
-    ind0 = 0
-    ind1 = g_0.num_faces * proj_dim
-    assert np.sum(subdomain_to_boundary[:, ind0:ind1]) == 8 * proj_dim
-    # Check that the projection matrix for the first fracture is non-zero. Since the
-    # fracture touches the boundary on two sides, we expect two non-zero rows.
-    ind0 = ind1
-    ind1 += g_1.num_faces * proj_dim
-    assert np.sum(subdomain_to_boundary[:, ind0:ind1]) == 2 * proj_dim
-    # Check that the projection matrix for the second fracture is non-zero.
-    ind0 = ind1
-    ind1 += g_2.num_faces * proj_dim
-    assert np.sum(subdomain_to_boundary[:, ind0:ind1]) == 2 * proj_dim
-    # The projection matrix for the intersection should be zero.
-    ind0 = ind1
-    assert np.sum(subdomain_to_boundary[:, ind0:]) == 0
-
-    # Make second projection on subset of grids.
-    subdomains = [g_0, g_1]
-    projection = pp.ad.grid_operators.BoundaryProjection(mdg, subdomains, proj_dim)
-    num_faces = proj_dim * (g_0.num_faces + g_1.num_faces)
-    num_cells = proj_dim * sum(
-        [mdg.subdomain_to_boundary_grid(sd).num_cells for sd in subdomains]
-    )
-    # Obtaining sparse matrices from the AD Operators.
-    subdomain_to_boundary = projection.subdomain_to_boundary.parse(mdg)
-    boundary_to_subdomain = projection.boundary_to_subdomain.parse(mdg)
-    # Check sizes.
-    assert subdomain_to_boundary.shape == (num_cells, num_faces)
-    assert boundary_to_subdomain.shape == (num_faces, num_cells)
-
-    # Check that the projection matrix for the top-dimensional grid is non-zero.
-    # Same sizes as above.
-    ind0 = 0
-    ind1 = g_0.num_faces * proj_dim
-    assert np.sum(subdomain_to_boundary[:, ind0:ind1]) == 8 * proj_dim
-    ind0 = ind1
-    ind1 += g_1.num_faces * proj_dim
-    assert np.sum(subdomain_to_boundary[:, ind0:ind1]) == 2 * proj_dim
-
-    # Check that subdomain_to_boundary and boundary_to_subdomain are consistent.
-    assert np.allclose((subdomain_to_boundary - boundary_to_subdomain.T).data, 0)
-
-
-@pytest.mark.integtest
-# Geometry based operators
-def test_trace(mdg: pp.MixedDimensionalGrid):
-    """Test Trace operator.
-
-    Parameters:
-        mdg: Mixed-dimensional grid.
-
-    This test is not ideal. It follows the implementation of Trace relatively closely,
-    but nevertheless provides some coverage, especially if Trace is carelessly changed.
-    The test constructs the expected md trace and inv_trace matrices and compares them
-    to the ones of Trace. Also checks that an error is raised if a non-scalar trace is
-    constructed (not implemented).
-    """
-    # The operator should work on any subset of mdg.subdomains.
-    subdomains = mdg.subdomains(dim=1)
-
-    # Construct expected matrices
-    traces, inv_traces = list(), list()
-    # No check on this function here.
-    # TODO: A separate unit test might be appropriate.
-    cell_projections, face_projections = pp.ad.grid_operators._subgrid_projections(
-        subdomains, dim=1
-    )
-    for sd in subdomains:
-        local_block = np.abs(sd.cell_faces.tocsr())
-        traces.append(local_block * cell_projections[sd].T)
-        inv_traces.append(local_block.T * face_projections[sd].T)
-
-    # Compare to operator class
-    op = pp.ad.Trace(subdomains)
-    _compare_matrices(op.trace, sps.bmat([[m] for m in traces]))
-    _compare_matrices(op.inv_trace, sps.bmat([[m] for m in inv_traces]))
-
-    # As of the writing of this test, Trace is not implemented for vector values.
-    # If it is ever extended, the test should be extended accordingly (e.g. parametrized with
-    # dim=[1, 2]).
-    with pytest.raises(NotImplementedError):
-        pp.ad.Trace(subdomains, dim=2)
-
-
-@pytest.mark.integtest
-@pytest.mark.parametrize("dim", [1, 4])
-def test_divergence(mdg: pp.MixedDimensionalGrid, dim: int):
-    """Test Divergence.
-
-    Parameters:
-        mdg: Mixed-dimensional grid.
-        dim: Dimension of vector field to which Divergence is applied.
-
-    This test is not ideal. It follows the implementation of Divergence relatively
-    closely, but nevertheless provides some coverage. Frankly, there is not much more to
-    do than comparing against the expected matrices, unless one wants to add more
-    integration-type tests e.g. evaluating combinations with other ad entities.
-
-    """
-    # The operator should work on any subset of mdg.subdomains.
-    subdomains = mdg.subdomains(dim=2) + mdg.subdomains(dim=0)
-
-    # Construct expected matrix
-    divergences = list()
-    for sd in subdomains:
-        # Kron does no harm if dim=1
-        local_block = sps.kron(sd.cell_faces.tocsr().T, sps.eye(dim))
-        divergences.append(local_block)
-
-    # Compare to operators parsed value
-    op = pp.ad.Divergence(subdomains)
-    val = op.parse(mdg)
-    _compare_matrices(val, sps.block_diag(divergences))
-
-
 @pytest.mark.integtest
 def test_ad_discretization_class():
     # Test of the parent class for all AD discretizations (pp.ad.Discretization)
@@ -1394,28 +1011,6 @@ def test_ad_discretization_class():
     assert np.allclose(known_val, discr_ad.foobar().parse(mdg).diagonal())
     assert np.allclose(known_sub_val, sub_discr_ad.foobar().parse(mdg).diagonal())
 
-
-## Below are helpers for tests of the Ad wrappers.
-def _compare_matrices(m1, m2):
-    if isinstance(m1, pp.ad.SparseArray):
-        m1 = m1._mat
-    if isinstance(m2, pp.ad.SparseArray):
-        m2 = m2._mat
-    if m1.shape != m2.shape:
-        return False
-    d = m1 - m2
-    if d.data.size > 0:
-        if np.max(np.abs(d.data)) > 1e-10:
-            return False
-    return True
-
-
-def _list_ind_of_grid(subdomains, g):
-    for i, gl in enumerate(subdomains):
-        if g == gl:
-            return i
-
-    raise ValueError("grid is not in list")
 
 
 class _MockDiscretization:
